+++ Nodify/Helpers/UnscaleTransformConverter.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Globalization;
+using System.Windows;
 using System.Windows.Data;
 using System.Windows.Media;
 
@@ -20,13 +21,27 @@ namespace Nodify
         }
     }
 
-    internal class UnscaleDoubleConverter : IMultiValueConverter
+    internal class ScaleDoubleConverter : IMultiValueConverter
     {
         public object Convert(IList<object?> values, Type targetType, object parameter, CultureInfo culture)
         {
             if (values.Count == 2 && values[0] is double d1 && values[1] is double d2)
                 return d1 * d2;
             return null;
+        }
+
+        public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
+    internal class ScalePointConverter : IMultiValueConverter
+    {
+        public object? Convert(IList<object?> values, Type targetType, object? parameter, CultureInfo culture)
+        {
+            Point result = (Point)((Vector)(Point)values[0] * (double)values[1]);
+            return result;
         }
 
         public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)